requirements.txt
Flask==2.2.3
mysql-connector-python==8.0.31
python-dotenv==0.21.0
